==> simulations/notebooks_sim_cts/0.3_sim_indepdent_rf_cts.ipynb <==
### summarize random forests results on Independent Simulation Scenarios for continuous outcome

dir = '/panfs/panfs1.ucsd.edu/panscratch/lij014/Stability_2020/sim_data'

load(paste0(dir, '/independent_RF.RData'))

dim.list = list()
size = c(50, 100, 500, 1000)
idx = 0
for (P in size){
    for (N in size){
        idx = idx + 1
        dim.list[[idx]] = c(P=P, N=N)
    }
}

files = NULL
for (dim in dim.list){
    p = dim[1]
    n = dim[2]
    files = cbind(files, paste0(dir, '/sim_independent_', paste('P', p, 'N', n, sep='_'), '.RData'))
}

avg_FDR = NULL
for (i in 1:length(files)){
    sim_file = files[i]
    load(sim_file, dat <- new.env())
    sub = dat$sim_array[[i]]
    p = sub$p # take true values from 1st replicate of each simulated data
    coef = sub$beta
    coef.true = which(coef != 0)
    
    tt = results_ind_rf[[i]]$Stab.table
    
    FDR = NULL
    for (r in 1:nrow(tt)){
        FDR = c(FDR, length(setdiff(which(tt[r, ] !=0), coef.true))/sum(tt[r, ]))

    }
    
    avg_FDR = c(avg_FDR, mean(FDR, na.rm=T))
}

table_ind = NULL
tmp_num_select = rep(0, length(results_ind_rf))
for (i in 1:length(results_ind_rf)){
    results_ind_rf[[i]]$OOB = paste(round(mean(results_ind_rf[[i]]$OOB.list, na.rm=T),2),
                                    '(', round(FSA::se(results_ind_rf[[i]]$OOB.list, na.rm=T),2), ')')
    table_ind = rbind(table_ind, results_ind_rf[[i]][c('n', 'p', 'rou', 'FP', 'FN', 'MSE', 'Stab', 'OOB')])
    tmp_num_select[i] = mean(rowSums(results_ind_rf[[i]]$Stab.table))
}
table_ind = as.data.frame(table_ind)
table_ind$num_select = tmp_num_select
table_ind$FDR = round(avg_FDR,2)

head(table_ind)

# export result
result.table_ind <- apply(table_ind,2,as.character)
rownames(result.table_ind) = rownames(table_ind)
result.table_ind = as.data.frame(result.table_ind)

# extract numbers only for 'n' & 'p'
result.table_ind$n = tidyr::extract_numeric(result.table_ind$n)
result.table_ind$p = tidyr::extract_numeric(result.table_ind$p)
result.table_ind$ratio = result.table_ind$p / result.table_ind$n

result.table_ind = result.table_ind[c('n', 'p', 'ratio', 'Stab', 'MSE', 'FP', 'FN', 'OOB', 'num_select', 'FDR')]
colnames(result.table_ind)[1:3] = c('N', 'P', 'Ratio')
result.table_ind$Index = seq(1, length(results_ind_rf), 1)

# convert interested measurements to be numeric
result.table_ind$Stab = as.numeric(as.character(result.table_ind$Stab))
result.table_ind$MSE_mean = as.numeric(substr(result.table_ind$MSE, start=1, stop=4))
result.table_ind$FP_mean = as.numeric(substr(result.table_ind$FP, start=1, stop=4))
result.table_ind$FN_mean = as.numeric(substr(result.table_ind$FN, start=1, stop=4))
result.table_ind$OOB_mean = as.numeric(substr(result.table_ind$OOB, start=1, stop=4))
result.table_ind$num_select = as.numeric(as.character(result.table_ind$num_select))
#result.table_ind$FDR = as.numeric(as.character(result.table_ind$FDR))

result.table_ind

## nothing was selected when Stab is NaN
for (i in c(1,5,9,13)){
    print(sum(results_ind_rf[[i]]$Stab.table))
}

## assign zero values to NA Stab; also recover missing FP values (1 in this case); missing FN values (6 in this case)
result.table_ind$Stab[is.na(result.table_ind$Stab)] = 0
result.table_ind$FP_mean[is.na(result.table_ind$FP_mean)] = 1
result.table_ind$FN_mean[is.na(result.table_ind$FN_mean)] = 6
#result.table_ind$FDR[is.na(result.table_ind$FDR)] = 0
result.table_ind

## export
write.table(result.table_ind, '../results_summary_cts/sim_ind_rf.txt', sep='\t', row.names=F)
