==> simulations/notebooks_sim_cts/1.3_sim_toe_rf_cts.ipynb <==
### summarize random forests results on Toeplitz Simulation Scenarios for continuous outcome

dir = '/panfs/panfs1.ucsd.edu/panscratch/lij014/Stability_2020/sim_data'

load(paste0(dir, '/toe_RF.RData'))

dim.list = list()
size = c(50, 100, 500, 1000)
idx = 0
for (P in size){
    for (N in size){
        idx = idx + 1
        dim.list[[idx]] = c(P=P, N=N)
    }
}

rou.list = seq(0.1, 0.9, 0.2)

files = NULL
for (rou in rou.list){
  for (dim in dim.list){
    p = dim[1]
    n = dim[2]
    files = cbind(files, paste0(dir, '/sim_toeplitz_corr', rou, paste('P', p, 'N', n, sep='_'), '.RData', sep=''))
  }
}

avg_FDR = NULL
for (i in 1:length(files)){
    sim_file = files[i]
    load(sim_file, dat <- new.env())
    sub = dat$sim_array[[i]]
    p = sub$p # take true values from 1st replicate of each simulated data
    coef = sub$beta
    coef.true = which(coef != 0)
    
    tt = results_toe_rf[[i]]$Stab.table
    
    FDR = NULL
    for (r in 1:nrow(tt)){
        FDR = c(FDR, length(setdiff(which(tt[r, ] !=0), coef.true))/sum(tt[r, ]))

    }
    
    avg_FDR = c(avg_FDR, mean(FDR, na.rm=T))
}

table_toe = NULL
tmp_num_select = rep(0, length(results_toe_rf))
for (i in 1:length(results_toe_rf)){
    results_toe_rf[[i]]$OOB = paste(round(mean(results_toe_rf[[i]]$OOB.list, na.rm=T),2),
                                    '(', round(FSA::se(results_toe_rf[[i]]$OOB.list, na.rm=T),2), ')')
    table_toe = rbind(table_toe, results_toe_rf[[i]][c('n', 'p', 'rou', 'FP', 'FN', 'MSE', 'Stab', 'OOB')])
    tmp_num_select[i] = mean(rowSums(results_toe_rf[[i]]$Stab.table))
}
table_toe = as.data.frame(table_toe)
table_toe$num_select = tmp_num_select
table_toe$FDR = round(avg_FDR,2)

head(table_toe)

# export result
result.table_toe <- apply(table_toe,2,as.character)
rownames(result.table_toe) = rownames(table_toe)
result.table_toe = as.data.frame(result.table_toe)

# extract numbers only for 'n' & 'p'
result.table_toe$n = tidyr::extract_numeric(result.table_toe$n)
result.table_toe$p = tidyr::extract_numeric(result.table_toe$p)
result.table_toe$ratio = result.table_toe$p / result.table_toe$n

result.table_toe = result.table_toe[c('n', 'p', 'rou', 'ratio', 'Stab', 'MSE', 'FP', 'FN', 'OOB', 'num_select', 'FDR')]
colnames(result.table_toe)[1:4] = c('N', 'P', 'Corr', 'Ratio')

head(result.table_toe)

# convert interested measurements to be numeric
result.table_toe$Stab = as.numeric(as.character(result.table_toe$Stab))
result.table_toe$MSE_mean = as.numeric(substr(result.table_toe$MSE, start=1, stop=4))
result.table_toe$FP_mean = as.numeric(substr(result.table_toe$FP, start=1, stop=4))
result.table_toe$FN_mean = as.numeric(substr(result.table_toe$FN, start=1, stop=4))
result.table_toe$FN_mean[is.na(result.table_toe$FN_mean)] = 0
result.table_toe$OOB_mean = as.numeric(substr(result.table_toe$OOB, start=1, stop=4))
result.table_toe$num_select = as.numeric(as.character(result.table_toe$num_select))

# check whether missing values exists
result.table_toe[rowSums(is.na(result.table_toe)) > 0,]

# recover values
result.table_toe$Stab[is.na(result.table_toe$Stab)] = 0
result.table_toe$MSE_mean[is.na(result.table_toe$MSE_mean)] = 1
result.table_toe$FP_mean[is.na(result.table_toe$FP_mean)] = 1
result.table_toe$FN_mean[result.table_toe$num_select == 0] = 6
result.table_toe$OOB_mean[is.na(result.table_toe$OOB_mean)] = 1

result.table_toe[c(1,5,9,13,14,17,21,25,29,33,37,41,45,49,53,57,61,65,69,73,77), ]

head(result.table_toe)

tail(result.table_toe)

## export
write.table(result.table_toe, '../results_summary_cts/sim_toe_rf.txt', sep='\t', row.names=F)
