==> simulations/notebooks_sim_cts/2.3_sim_block_rf_cts.ipynb <==
### summarize random forests results on Block Simulation Scenarios for continuous outcome

dir = '/panfs/panfs1.ucsd.edu/panscratch/lij014/Stability_2020/sim_data'

load(paste0(dir, '/block_RF.RData'))

dim.list = list()
size = c(50, 100, 500, 1000)
idx = 0
for (P in size){
    for (N in size){
        idx = idx + 1
        dim.list[[idx]] = c(P=P, N=N)
    }
}

rou.list = seq(0.1, 0.9, 0.2)

files = NULL
for (rou in rou.list){
  for (dim in dim.list){
    p = dim[1]
    n = dim[2]
    files = cbind(files, paste0(dir, '/sim_block_corr', rou, paste('P', p, 'N', n, sep='_'), '.RData', sep=''))
  }
}

avg_FDR = NULL
for (i in 1:length(files)){
    sim_file = files[i]
    load(sim_file, dat <- new.env())
    sub = dat$sim_array[[i]]
    p = sub$p # take true values from 1st replicate of each simulated data
    coef = sub$beta
    coef.true = which(coef != 0)
    
    tt = results_block_rf[[i]]$Stab.table
    
    FDR = NULL
    for (r in 1:nrow(tt)){
        FDR = c(FDR, length(setdiff(which(tt[r, ] !=0), coef.true))/sum(tt[r, ]))

    }
    
    avg_FDR = c(avg_FDR, mean(FDR, na.rm=T))
}

table_block = NULL
tmp_num_select = rep(0, length(results_block_rf))
for (i in 1:length(results_block_rf)){
    results_block_rf[[i]]$OOB = paste(round(mean(results_block_rf[[i]]$OOB.list, na.rm=T),2),
                                    '(', round(FSA::se(results_block_rf[[i]]$OOB.list, na.rm=T),2), ')')
    table_block = rbind(table_block, results_block_rf[[i]][c('n', 'p', 'rou', 'FP', 'FN', 'MSE', 'Stab', 'OOB')])
    tmp_num_select[i] = mean(rowSums(results_block_rf[[i]]$Stab.table))
}
table_block = as.data.frame(table_block)
table_block$num_select = tmp_num_select
table_block$FDR = round(avg_FDR,2)

head(table_block)

# export result
result.table_block <- apply(table_block,2,as.character)
rownames(result.table_block) = rownames(table_block)
result.table_block = as.data.frame(result.table_block)

# extract numbers only for 'n' & 'p'
result.table_block$n = tidyr::extract_numeric(result.table_block$n)
result.table_block$p = tidyr::extract_numeric(result.table_block$p)
result.table_block$ratio = result.table_block$p / result.table_block$n

result.table_block = result.table_block[c('n', 'p', 'rou', 'ratio', 'Stab', 'MSE', 'FP', 'FN', 'OOB', 'num_select', 'FDR')]
colnames(result.table_block)[1:4] = c('N', 'P', 'Corr', 'Ratio')

# convert interested measurements to be numeric
result.table_block$Stab = as.numeric(as.character(result.table_block$Stab))
result.table_block$MSE_mean = as.numeric(substr(result.table_block$MSE, start=1, stop=4))
result.table_block$FP_mean = as.numeric(substr(result.table_block$FP, start=1, stop=4))
result.table_block$FN_mean = as.numeric(substr(result.table_block$FN, start=1, stop=4))
result.table_block$FN_mean[is.na(result.table_block$FN_mean)] = 0
result.table_block$OOB_mean = as.numeric(substr(result.table_block$OOB, start=1, stop=4))
result.table_block$num_select = as.numeric(as.character(result.table_block$num_select))

# check whether missing values exists
result.table_block[rowSums(is.na(result.table_block)) > 0,]

# recover values
result.table_block$Stab[is.na(result.table_block$Stab)] = 0
result.table_block$FP_mean[is.na(result.table_block$FP_mean)] = 1
result.table_block$OOB_mean[is.na(result.table_block$OOB_mean)] = 2
result.table_block$FN_mean[result.table_block$num_select == 0] = 6

result.table_block[c(1,5,9,13,17,21,25,29,33,37,41,45,49,53,57,61,65,69,73,77), ]

head(result.table_block)

tail(result.table_block)

## export
write.table(result.table_block, '../results_summary_cts/sim_block_rf.txt', sep='\t', row.names=F)
