==> simulations/notebooks_sim_cts/3_sim_cts_summary.ipynb <==
### summarize method comparison results across 4 feature selection methods in different simulation scenarios

### 1. Independent Design

ind_lasso = read.csv('../results_summary_cts/sim_ind_lasso.txt', sep='\t')
ind_elnet = read.csv('../results_summary_cts/sim_ind_elnet.txt', sep='\t')
ind_rf = read.csv('../results_summary_cts/sim_ind_rf.txt', sep='\t')
ind_compLasso = read.csv('../results_summary_cts/sim_ind_compLasso.txt', sep='\t')

ind_lasso$method = rep('lasso', dim(ind_lasso)[1])
ind_elnet$method = rep('elnet', dim(ind_elnet)[1])
ind_rf$method = rep('rf', dim(ind_rf)[1])
ind_compLasso$method = rep('compLasso', dim(ind_compLasso)[1])

dim(ind_lasso)
dim(ind_elnet)
dim(ind_rf)
dim(ind_compLasso)

ind_rf = ind_rf[!names(ind_rf)%in% c('OOB', 'OOB_mean', 'Index')]
dim(ind_rf)

ind = rbind(ind_lasso, ind_elnet)
ind = rbind(ind, ind_rf)
ind = rbind(ind, ind_compLasso)
dim(ind)

# fix false positives: when nothing was selected, then fp should be zero
# fp is now 1 due to length(setdiff(0/NA)) = 1 when actually should be 0 (in cv_sim_apply.R)
ind$FP_mean[ind$num_select ==0] = 0

write.table(ind, '../results_summary_cts/table_ind_all.txt', sep='\t', row.names=F)

ind[order(ind$N, ind$P, decreasing=T), c('N', 'P', 'method', 'Stab', 'MSE_mean', 'FP_mean', 'FN_mean', 'num_select', 'FDR')][1:20, ]

### 2. Toeplitz Design

toe_lasso = read.csv('../results_summary_cts/sim_toe_lasso.txt', sep='\t')
toe_elnet = read.csv('../results_summary_cts/sim_toe_elnet.txt', sep='\t')
toe_rf = read.csv('../results_summary_cts/sim_toe_rf.txt', sep='\t')
toe_compLasso = read.csv('../results_summary_cts/sim_toe_compLasso.txt', sep='\t')

toe_lasso$method = rep('lasso', dim(toe_lasso)[1])
toe_elnet$method = rep('elnet', dim(toe_elnet)[1])
toe_rf$method = rep('rf', dim(toe_rf)[1])
toe_compLasso$method = rep('compLasso', dim(toe_compLasso)[1])

dim(toe_lasso)
dim(toe_elnet)
dim(toe_rf)
dim(toe_compLasso)

toe_rf = toe_rf[!names(toe_rf)%in% c('OOB', 'OOB_mean', 'toeex')]
dim(toe_rf)

toe = rbind(toe_lasso, toe_elnet)
toe = rbind(toe, toe_rf)
toe = rbind(toe, toe_compLasso)
dim(toe)

toe[toe$num_select == 0, ]

# fix false positives: when nothing was selected, then fp should be zero
# fp is now 1 due to length(setdiff(0/NA)) = 1 when actually should be 0 (in cv_sim_apply.R)
toe$FP_mean[toe$num_select ==0] = 0
write.table(toe, '../results_summary_cts/table_toe_all.txt', sep='\t', row.names=F)

### 3. Block Design

block_lasso = read.csv('../results_summary_cts/sim_block_lasso.txt', sep='\t')
block_elnet = read.csv('../results_summary_cts/sim_block_elnet.txt', sep='\t')
block_rf = read.csv('../results_summary_cts/sim_block_rf.txt', sep='\t')
block_compLasso = read.csv('../results_summary_cts/sim_block_compLasso.txt', sep='\t')

block_lasso$method = rep('lasso', dim(block_lasso)[1])
block_elnet$method = rep('elnet', dim(block_elnet)[1])
block_rf$method = rep('rf', dim(block_rf)[1])
block_compLasso$method = rep('compLasso', dim(block_compLasso)[1])

dim(block_lasso)
dim(block_elnet)
dim(block_rf)
dim(block_compLasso)

block_rf = block_rf[!names(block_rf)%in% c('OOB', 'OOB_mean', 'blockex')]
dim(block_rf)

block = rbind(block_lasso, block_elnet)
block = rbind(block, block_rf)
block = rbind(block, block_compLasso)
dim(block)

# fix false positives: when nothing was selected, then fp should be zero
# fp is now 1 due to length(setdiff(0/NA)) = 1 when actually should be 0 (in cv_sim_apply.R)
block$FP_mean[block$num_select ==0] = 0
write.table(block, '../results_summary_cts/table_block_all.txt', sep='\t', row.names=F)
