==> simulations/notebooks_sim_cts/4_sim_cts_bootstrap_testing.ipynb <==
### Hypothesis testing based on boostrap for selected simulation scenario

dir = '/panfs/panfs1.ucsd.edu/panscratch/lij014/Stability_2020/sim_data'

load(paste0(dir, '/boot_toe_RF.RData'))
load(paste0(dir, '/boot_block_RF.RData'))
load(paste0(dir, '/boot_toe_compLasso.RData'))
load(paste0(dir, '/boot_block_compLasso.RData'))

table = as.data.frame(cbind(toe_rf$stab_index, toe_lin$stab_index, block_rf$stab_index, block_lin$stab_index))
colnames(table) = c('toe_rf', 'toe_lin', 'block_rf', 'block_lin')
head(table)

## Two sample test based on Bootstrap

## interpretation: if confidence interval doesn't contain zero, then significant
## ref: correct intrepretation of confidence intervals: https://journals.sagepub.com/doi/pdf/10.1177/201010581001900316

diff_toe = (table$toe_lin - table$toe_rf)
mean(diff_toe)
quantile(diff_toe, probs = c(0.025, 0.975))

diff_block = (table$block_lin - table$block_rf)
mean(diff_block)
quantile(diff_block, probs = c(0.025, 0.975))

## Compare based on MSE

#### block 0.5

table_block_rf = NULL
for (i in 1:length(results_block_rf)){
    table_block_rf = rbind(table_block_rf, results_block_rf[[i]][c('n', 'p', 'rou', 'MSE', 'MSE.list')])
}
table_block_rf = as.data.frame(table_block_rf)
rf_block5_rf= table_block_rf[table_block_rf$n == 100 & table_block_rf$p == 1000 & table_block_rf$rou == 0.5, ]
rf_block5_rf

table_block_compLasso = NULL
for (i in 1:length(results_block_compLasso)){
    table_block_compLasso = rbind(table_block_compLasso, results_block_compLasso[[i]][c('n', 'p', 'rou', 'MSE', 'MSE.list')])
}
table_block_compLasso = as.data.frame(table_block_compLasso)
rf_block5_compLasso= table_block_compLasso[table_block_compLasso$n == 100 & table_block_compLasso$p == 1000 & table_block_compLasso$rou == 0.5, ]
rf_block5_compLasso

diff_block = (unlist(rf_block5_compLasso$MSE.list) - unlist(rf_block5_rf$MSE.list))
mean(diff_block)
sd(diff_block)
quantile(diff_block, probs = c(0.025, 0.975))

### Toeplitz 0.5

table_toe_rf = NULL
for (i in 1:length(results_toe_rf)){
    table_toe_rf = rbind(table_toe_rf, results_toe_rf[[i]][c('n', 'p', 'rou', 'MSE', 'MSE.list')])
}
table_toe_rf = as.data.frame(table_toe_rf)
rf_toe5_rf= table_toe_rf[table_toe_rf$n == 100 & table_toe_rf$p == 1000 & table_toe_rf$rou == 0.5, ]
rf_toe5_rf

table_toe_compLasso = NULL
for (i in 1:length(results_toe_compLasso)){
    table_toe_compLasso = rbind(table_toe_compLasso, results_toe_compLasso[[i]][c('n', 'p', 'rou', 'MSE', 'MSE.list')])
}
table_toe_compLasso = as.data.frame(table_toe_compLasso)
rf_toe5_compLasso= table_toe_compLasso[table_toe_compLasso$n == 100 & table_toe_compLasso$p == 1000 & table_toe_compLasso$rou == 0.5, ]
rf_toe5_compLasso

diff_toe = (unlist(rf_toe5_compLasso$MSE.list) - unlist(rf_toe5_rf$MSE.list))
mean(diff_toe)
sd(diff_toe)
quantile(diff_toe, probs = c(0.025, 0.975))
